% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{families_Bf}
\alias{families_Bf}
\title{Gene families in \eqn{Bacteroides fragilis} 3_1_12}
\format{A character vector.}
\source{
STRING 10.0 database \url{http://string-db.org/}.
}
\usage{
families_Bf
}
\description{
List of gene families present in \eqn{Bacteroides fragilis} 3_1_12 as annotated in STRING 10.0.
}
\examples{
data(families_Bf)
}
\references{
\enumerate{
           \item Szklarczyk D, Franceschini A, Wyder S, Forslund K, Heller D, Huerta-Cepas J,
           Simonovic M, Roth A, Santos A, Tsafou KP, Kuhn M, Bork P, Jensen LJ, von Mering C.
           STRING v10: protein-protein interaction networks, integrated over the tree of life.
           Nucleic Acids Res. 2015 Jan; 43:D447-52.4.
         }
}
\keyword{datasets}

