% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_associations.R
\name{functional_association}
\alias{functional_association}
\title{Compute microbe-microbe functional association index}
\usage{
functional_association(families1, families2, network = reference_network)
}
\arguments{
\item{families1}{character vector of gene family names from STRING 10.0 present in first genome.}

\item{families2}{character vector of gene family names from STRING 10.0 present in second genome.}

\item{network}{data frame encoding associations between gene families. Should contain 2 or 3 variables:
\code{node1} gene family one, \code{node2} gene family two,
\code{weight} weight of the edge.}
}
\value{
Microbe-microbe functional association index for two genomes which encode \code{families1} and \code{families1}.
       This index represents fraction of edges which connect gene families present in one of the compared genomes to gene families present in another
       out of all of the edges in the gene family network encoded by one of the two genomes exclusively.
}
\description{
This function calculates microbe-microbe functional association index for two genomes.
}
\examples{
data(families_Bf)
data(families_Er)
association_index = functional_association(families1 = families_Bf, families2 = families_Er, network = reference_network)
association_index
}
\references{
\enumerate{
           \item Kamneva OK. Genome composition of microbes predicts their co-occurrence in the environment. In review.
         }
}
\seealso{
\code{\link{similarity}}, \code{\link{similarity_by_set}}.
}

