% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics2ecology-package.R
\docType{package}
\name{genomics2ecology-package}
\alias{genomics2ecology-package}
\title{genomics2ecology: Package for genomic analysis for microbial ecology.}
\description{
The genomics2ecology package provides two categories of functions
             which compute genome content similarity or microbe-microbe association indices
             for a pair of genomes using provided a list of STRING 10.0 gene families present in each
             of the genome.\cr \cr
             The complete list of functions can be displayed with \code{library(help = genomics2ecology)}.
}
\author{
Olga K. Kamneva: <olga dot k dot kamneva at gmail dot com>
}

