% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_gene_sets}
\alias{reference_gene_sets}
\title{Reference gene sets}
\format{List of character vectors. Each individual vector represents a set of functionally associated gene families (or gene set)
       likely involved in the same function or process.}
\usage{
reference_gene_sets
}
\description{
Sets of functionally linked gene families identified using MCL clustering of the global network
             (\code{reference_network}) of all the gene families identified in STRING 10.0 database.
}
\examples{
data(reference_gene_sets)
}
\references{
\enumerate{
           \item Kamneva OK. Genome composition of microbes predicts their co-occurrence in the environment. In review.
         }
}
\keyword{datasets}

