% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_representation.R
\name{set_representation}
\alias{set_representation}
\title{Calculate gene set representation}
\usage{
set_representation(families, gene_sets = reference_gene_sets)
}
\arguments{
\item{families}{character vector of gene family names from STRING 10.0 present in first genome.}

\item{gene_sets}{list of character vectors. Each individual vector represents a set of functionally associated gene families (or gene set)
likely involved in the same function or process.}
}
\value{
Microbe-microbe functional association index for two genomes which encode \code{families1} and \code{families1}.
}
\description{
This function calculates presence and absence of gene families in each of the reference gene set in a genome from
             the list of gene family encoded in that genome.
}
\examples{
data(families_Bf)
sets_Bf = set_representation(families = families_Bf, gene_sets = reference_gene_sets)
}
\references{
\enumerate{
           \item Kamneva OK. Genome composition of microbes predicts their co-occurrence in the environment. In review.
         }
}
\seealso{
\code{\link{similarity}}, \code{\link{similarity_by_set}}.
}

