% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_content.R
\name{similarity}
\alias{similarity}
\title{Compute genome content similarity index}
\usage{
similarity(gene_sets1, gene_sets2, threshold = 0.05, size = 4)
}
\arguments{
\item{gene_sets1}{list of vectors of 1s and 0s which represnet presence and absence of gene families from each of the gene sets in first genome.}

\item{gene_sets2}{list of vectors of 1s and 0s which represnet presence and absence of gene families from each of the gene sets in second genome.}

\item{threshold}{fraction of gene families from a gene set which is required to be present in
at least one of the genomes for for the gene set to contribute to genome content
similarity index calculation (default is 0.05).}

\item{size}{minimal size of the gene set to consider (default is 4).}
}
\value{
Genome content similarity index computed for two genomes encoding \code{gene_sets1}
       and \code{gene_sets2} respectively.
}
\description{
This function calculates genome content similarity index for two genomes.
}
\examples{
data(families_Bf)
data(families_Er)
sets_Bf = set_representation(families = families_Bf, gene_sets = reference_gene_sets)
sets_Er = set_representation(families = families_Er, gene_sets = reference_gene_sets)
sim_Bf_Er = similarity(gene_sets1 = sets_Bf, gene_sets2 = sets_Er, threshold = 0.05, size = 4)
sim_Bf_Er
}
\references{
\enumerate{
           \item Kamneva OK. Genome composition of microbes predicts their co-occurrence in the environment. In review.
         }
}
\seealso{
\code{\link{similarity_by_set}}, \code{\link{functional_association}}, \code{\link{set_representation}},
         \code{\link{reference_gene_sets}}.
}

